requirements/base.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
django-environ
psycopg2-binary
django-storages
boto3
